+++ components/Viewer3D.tsx
@@ -1,15 +1,17 @@
 "use client"
 
 import { Canvas, useThree, useFrame } from "@react-three/fiber"
-import { OrbitControls, useGLTF, Text } from "@react-three/drei"
+import { useGLTF, Text } from "@react-three/drei"
 import { CustomGrid } from "./CustomGrid"
 import { useEffect, useRef, useState, useMemo } from "react"
 import * as THREE from "three"
-import { plyAsyncParse } from "@/utils/ply-parser.web"
-import type { OrbitControls as OrbitControlsImpl } from 'three-stdlib'
 import type { Portal } from "@/utils/posemeshClientApi"
 import { matrixFromPose } from "@/utils/three-utils"
-import { OBJLoader } from 'three/examples/jsm/loaders/OBJLoader.js'
+import { Portals } from "./3d/Portals"
+import { OcclusionMesh } from "./3d/OcclusionMesh"
+import { NavMesh } from "./3d/NavMesh"
+import { CameraController } from "./3d/CameraController"
+import { PointCloud } from "./3d/PointCloud"
 
 /**
  * Device data interface for displaying devices in 3D space
@@ -43,326 +45,6 @@ interface Viewer3DProps {
   pointCloudVisible?: boolean
   scan3DVisible?: boolean
   domainDeviceData?: DeviceData[] | null
-}
-
-/**
- * Renders a point cloud from PLY file data with vertex colors.
- * 
- * @param data - ArrayBuffer containing the PLY file data
- */
-function PointCloud({ data }: { data: ArrayBuffer }) {
-  const { scene } = useThree()
-  const pointsRef = useRef<THREE.Points | null>(null)
-
-  useEffect(() => {
-    if (!data) return
-
-    const cleanupRef = { current: () => {} }
-
-    plyAsyncParse(data, true).then((geometry) => {
-      const material = new THREE.PointsMaterial({
-        size: 0.09,
-        vertexColors: true,
-        sizeAttenuation: true,
-        depthWrite: true,
-        opacity: 1,
-        transparent: true,
-      })
-
-      const points = new THREE.Points(geometry, material)
-      scene.add(points)
-      pointsRef.current = points
-      
-      // Define cleanup for this effect
-      cleanupRef.current = () => {
-        scene.remove(points)
-        geometry.dispose()
-        material.dispose()
-      }
-    }).catch(error => {
-      console.error("Error parsing PLY data:", error)
-    })
-
-    return () => {
-      cleanupRef.current()
-    }
-  }, [data, scene])
-
-  return null
-}
-
-/**
- * Renders portal markers (QR codes) at specified positions and orientations.
- * Uses a 3D model loaded from QR.glb.
- * 
- * @param portals - Array of Portal objects containing position and orientation data
- */
-function Portals({ portals = [] }: { portals: Portal[] | null | undefined }) {
-  const { scene: gltfScene } = useGLTF('/QR.glb')
-  const { scene } = useThree()
-  const modelsRef = useRef<Map<string, THREE.Group>>(new Map())
-  const matrix = useMemo(() => new THREE.Matrix4(), [])
-
-  useEffect(() => {
-    if (!gltfScene || !portals || portals.length === 0) return
-
-    // Create or update models for each portal
-    portals.forEach(portal => {
-      let model: THREE.Group
-      
-      if (modelsRef.current.has(portal.id)) {
-        model = modelsRef.current.get(portal.id)!
-      } else {
-        model = gltfScene.clone()
-        scene.add(model)
-        modelsRef.current.set(portal.id, model)
-      }
-
-      // Set portal position and rotation
-      if (matrixFromPose(portal, matrix)) {
-        matrix.decompose(model.position, model.quaternion, model.scale)
-        
-        // Apply the reported size if available
-        if (portal.reported_size) {
-          const size = portal.reported_size * 0.01 // Convert to meters
-          model.scale.setScalar(size)
-        }
-      }
-    })
-
-    // Cleanup removed portals
-    modelsRef.current.forEach((model, id) => {
-      if (!portals.find(p => p.id === id)) {
-        scene.remove(model)
-        disposeModel(model)
-        modelsRef.current.delete(id)
-      }
-    })
-
-    // Cleanup function
-    return () => {
-      modelsRef.current.forEach((model) => {
-        scene.remove(model)
-        disposeModel(model)
-      })
-      modelsRef.current.clear()
-    }
-  }, [gltfScene, scene, portals, matrix])
-
-  return null
-}
-
-/**
- * Helper function to dispose of Three.js model resources
- */
-function disposeModel(model: THREE.Object3D) {
-  model.traverse((object) => {
-    if (object instanceof THREE.Mesh) {
-      if (object.geometry) {
-        object.geometry.dispose()
-      }
-      if (object.material instanceof THREE.Material) {
-        object.material.dispose()
-      } else if (Array.isArray(object.material)) {
-        object.material.forEach(material => material.dispose())
-      }
-    }
-  })
-}
-
-/**
- * Controls camera behavior including auto-rotation when idle.
- * 
- * @param pointCloudData - Point cloud data used to determine if content is loaded
- */
-function CameraController({ pointCloudData }: { pointCloudData: ArrayBuffer | null }) {
-  const { camera } = useThree()
-  const controlsRef = useRef<OrbitControlsImpl>(null)
-  const [isIdle, setIsIdle] = useState(false)
-  const lastInteractionTime = useRef(Date.now())
-  const animationRef = useRef<number | null>(null)
-  
-  // Time before starting auto-rotation (ms)
-  const IDLE_TIMEOUT = 500000
-
-  const resetIdleTimer = () => {
-    lastInteractionTime.current = Date.now()
-    if (isIdle) {
-      setIsIdle(false)
-    }
-  }
-
-  const startOrbitAnimation = () => {
-    if (controlsRef.current && !animationRef.current) {
-      const animate = () => {
-        if (controlsRef.current) {
-          controlsRef.current.autoRotate = true
-          controlsRef.current.update()
-        }
-        animationRef.current = requestAnimationFrame(animate)
-      }
-      animate()
-    }
-  }
-
-  const stopOrbitAnimation = () => {
-    if (animationRef.current) {
-      cancelAnimationFrame(animationRef.current)
-      animationRef.current = null
-    }
-    if (controlsRef.current) {
-      controlsRef.current.autoRotate = false
-    }
-  }
-
-  // Check for idle state in animation frame
-  useFrame(() => {
-    if (pointCloudData && !isIdle && 
-        Date.now() - lastInteractionTime.current > IDLE_TIMEOUT) {
-      setIsIdle(true)
-      startOrbitAnimation()
-    }
-  })
-
-  // Cleanup animation frame on unmount
-  useEffect(() => {
-    return () => {
-      if (animationRef.current) {
-        cancelAnimationFrame(animationRef.current)
-      }
-    }
-  }, [])
-
-  const handleStart = () => {
-    resetIdleTimer()
-    stopOrbitAnimation()
-  }
-
-  const handleEnd = () => {
-    resetIdleTimer()
-  }
-
-  return (
-    <OrbitControls
-      ref={controlsRef}
-      minPolarAngle={0}
-      maxPolarAngle={Math.PI / 2}
-      makeDefault
-      autoRotateSpeed={0.5}
-      enableDamping={true}
-      dampingFactor={0.05}
-      onStart={handleStart}
-      onEnd={handleEnd}
-      onChange={resetIdleTimer}
-    />
-  )
-}
-
-/**
- * Renders the occlusion mesh that represents physical barriers in the space.
- * 
- * @param occlusionMeshData - ArrayBuffer containing the OBJ file data
- */
-function OcclusionMesh({ occlusionMeshData }: { occlusionMeshData: ArrayBuffer | null }) {
-  const { scene } = useThree()
-  const groupRef = useRef<THREE.Group | null>(null)
-
-  useEffect(() => {
-    if (!occlusionMeshData) return
-
-    const loader = new OBJLoader()
-    const objString = new TextDecoder().decode(occlusionMeshData)
-    const obj = loader.parse(objString)
-    
-    // Create a group to hold all meshes
-    const group = new THREE.Group()
-    
-    // Process all children in the OBJ
-    obj.traverse((child) => {
-      if (child instanceof THREE.Mesh) {
-        // Create wireframe geometry
-        const wireframe = new THREE.WireframeGeometry(child.geometry)
-        const edges = new THREE.LineSegments(
-          wireframe,
-          new THREE.LineBasicMaterial({ color: 0x303030 })
-        )
-
-        // Create mesh with transparent faces
-        const mesh = new THREE.Mesh(
-          child.geometry,
-          new THREE.MeshBasicMaterial({ 
-            color: 0x808080,
-            transparent: true,
-            opacity: 0.8
-          })
-        )
-
-        group.add(mesh)
-        group.add(edges)
-      }
-    })
-
-    scene.add(group)
-    groupRef.current = group
-
-    return () => {
-      if (groupRef.current) {
-        scene.remove(group)
-        disposeModel(group)
-      }
-    }
-  }, [occlusionMeshData, scene])
-
-  return null
-}
-
-/**
- * Renders the navigation mesh that represents walkable areas in the space.
- * 
- * @param navMeshData - ArrayBuffer containing the OBJ file data
- */
-function NavMesh({ navMeshData }: { navMeshData: ArrayBuffer | null }) {
-  const { scene } = useThree()
-  const groupRef = useRef<THREE.Group | null>(null)
-
-  useEffect(() => {
-    if (!navMeshData) return
-
-    const loader = new OBJLoader()
-    const objString = new TextDecoder().decode(navMeshData)
-    const obj = loader.parse(objString)
-    
-    // Create a group to hold all meshes
-    const group = new THREE.Group()
-    
-    // Process all children in the OBJ
-    obj.traverse((child) => {
-      if (child instanceof THREE.Mesh) {
-        const mesh = new THREE.Mesh(
-          child.geometry,
-          new THREE.MeshBasicMaterial({ 
-            color: 0x2B4D2B,
-            transparent: true,
-            opacity: 0.9,
-            side: THREE.DoubleSide
-          })
-        )
-        group.add(mesh)
-      }
-    })
-
-    scene.add(group)
-    groupRef.current = group
-
-    return () => {
-      if (groupRef.current) {
-        scene.remove(group)
-        disposeModel(group)
-      }
-    }
-  }, [navMeshData, scene])
-
-  return null
 }
 
 /**
@@ -957,11 +639,27 @@ export default function Viewer3D({
   )
 }
 
-// Preload all GLTF models
-useGLTF.preload('/QR.glb')
 // useGLTF.preload('/L10CommonSpace.glb')
 // useGLTF.preload('/lounge.glb')
 useGLTF.preload('/glasses.glb')
 useGLTF.preload('/smartphone.glb')
 useGLTF.preload('/padbot-robot-w3.glb')
+
+/**
+ * Helper function to dispose of Three.js model resources
+ */
+function disposeModel(model: THREE.Object3D) {
+  model.traverse((object) => {
+    if (object instanceof THREE.Mesh) {
+      if (object.geometry) {
+        object.geometry.dispose()
+      }
+      if (object.material instanceof THREE.Material) {
+        object.material.dispose()
+      } else if (Array.isArray(object.material)) {
+        object.material.forEach(material => material.dispose())
+      }
+    }
+  })
+}
